Cleaned dockerfile and config of discovery service

+++ discovery_service/src/drop_database.py
@@ -1,5 +1,6 @@
 import logging
 
+import sqlalchemy_utils
 from sqlalchemy_utils import drop_database
 
 from .config import DB_CONN_STRING
@@ -8,8 +9,11 @@ logging.getLogger().setLevel(logging.INFO)
 
 
 def drop_db() -> None:
-    drop_database(DB_CONN_STRING)
-    logging.info(f'Services location database deleted')
+    if sqlalchemy_utils.database_exists(DB_CONN_STRING):
+        drop_database(DB_CONN_STRING)
+        logging.info('Services location database deleted')
+    else:
+        logging.info("Database does not exist.")
 
 
 if __name__ == '__main__':
